rewrite executor tests
